Rm tailwind, deduplicate heritability models
