+++ package.json
@@ -63,8 +63,6 @@
     "eslint": "^8.57.0",
     "eslint-config-next": "^14",
     "graphql": "^16",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.3",
     "typescript": "^5"
   }
 }
